Add files via upload
The codes of this project was written on Google's big query platform. This project reviewed Covid-19 data from 2020 to 2022. I was able to create a template to view the daily rolling sum and percentage of cases, deaths and vaccinations. The result displayed the gradual decline in cases from over 2% of world population to less than 1%.